requirements.txt
playwright>=1.40.0
python-dotenv>=1.0.0
Pillow>=10.0.0
flask>=3.0.0

# Server communication
httpx>=0.28.0

# Content generation (free AI API fallback chain)
google-genai>=1.0.0
mistralai>=1.0.0
groq>=0.4.0

# Image post-processing (UGC authenticity)
numpy>=1.24.0
piexif>=1.1.3

# Desktop experience
pystray>=0.19.0
plyer>=2.1.0
